movie/requirements.txt
Pillow==8.4.0
oandas==1.2.3